requirements.txt
chardet==3.0.4
jsonpatch==1.23
kubernetes==12.0.0a1
pymongo==3.7.1
PyYAML==5.1
requests==2.22.0
urllib3==1.25.3
dnspython==1.16.0
cryptography==2.7
python-box==5.3.0
